Update Auto_Message.py
It is very tiring to send messages to dozens of people at the same time on special days such as Kandil and Eid. To solve this problem, I updated the code so that I can send messages one after the other. The code sends messages at 1-minute intervals despite WhatsApp's spam blocks.